refactor: Distance 관련 클래스 수정
